Do not import sqlalchemy when running without database
